refactor and add generic matrix
